Dropping half, reworking to nightly f16

+++ src/neon/f16_utils.rs
@@ -41,9 +41,11 @@ pub struct x_float16x4_t(pub(crate) uint16x4_t);
 #[allow(dead_code)]
 pub struct x_float16x8_t(pub(crate) uint16x8_t);
 
+use core::f16;
+
 #[inline]
-pub unsafe fn xvld_f16(ptr: *const half::f16) -> x_float16x4_t {
-    let store: uint16x4_t = vld1_u16(std::mem::transmute::<*const half::f16, *const u16>(ptr));
+pub unsafe fn xvld_f16(ptr: *const f16) -> x_float16x4_t {
+    let store: uint16x4_t = vld1_u16(std::mem::transmute::<*const f16, *const u16>(ptr));
     std::mem::transmute(store)
 }
 
@@ -86,9 +88,9 @@ pub(super) unsafe fn xvcvt_f16_f32(v: float32x4_t) -> x_float16x4_t {
 }
 
 #[inline]
-pub unsafe fn xvst_f16(ptr: *mut half::f16, x: x_float16x4_t) {
+pub unsafe fn xvst_f16(ptr: *mut f16, x: x_float16x4_t) {
     vst1_u16(
-        std::mem::transmute::<*mut half::f16, *mut u16>(ptr),
+        std::mem::transmute::<*mut f16, *mut u16>(ptr),
         xreinterpret_u16_f16(x),
     )
 }
